api/agents/claims_processor/requirements.txt
python-docx
PyPDF2
requests
crewai
crewai-tools
python-dotenv